# api/openapi.json
{
  "components": {
    "schemas": {
      "ClusterListResponse": {
        "description": "Paginated cluster summaries.",
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "pagination",
          "items"
        ],
        "title": "ClusterListResponse",
        "type": "object"
      },
      "EvaluationResponse": {
        "description": "Read-only evaluation summary response.",
        "properties": {
          "summaries": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Summaries",
            "type": "object"
          }
        },
        "required": [
          "summaries"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "API readiness response.",
        "properties": {
          "artifact_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Artifact Counts",
            "type": "object"
          },
          "config_name": {
            "title": "Config Name",
            "type": "string"
          },
          "corpus": {
            "title": "Corpus",
            "type": "string"
          },
          "document_count": {
            "title": "Document Count",
            "type": "integer"
          },
          "loaded_search_run_id": {
            "title": "Loaded Search Run Id",
            "type": "string"
          },
          "readiness": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Readiness",
            "type": "object"
          },
          "status": {
            "enum": [
              "ok",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "status",
          "config_name",
          "corpus",
          "loaded_search_run_id",
          "document_count",
          "readiness",
          "artifact_counts",
          "warnings"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "Pagination": {
        "description": "Page-number pagination metadata.",
        "properties": {
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "page_size",
          "total"
        ],
        "title": "Pagination",
        "type": "object"
      },
      "RelationListResponse": {
        "description": "Paginated relation summaries.",
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "pagination",
          "items"
        ],
        "title": "RelationListResponse",
        "type": "object"
      },
      "SearchRequest": {
        "description": "Search request body.",
        "properties": {
          "page": {
            "default": 1,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "system": {
            "default": "fused",
            "enum": [
              "fused",
              "title_bm25",
              "all_vocabulary_bm25",
              "pearl_semantic"
            ],
            "title": "System",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
      "SearchResponse": {
        "description": "Grounded table-search response.",
        "properties": {
          "notice": {
            "title": "Notice",
            "type": "string"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "notice",
          "pagination",
          "results"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "TableDetailResponse": {
        "description": "Denormalized source-table detail.",
        "properties": {
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          },
          "geographies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Geographies",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "parse_status": {
            "title": "Parse Status",
            "type": "string"
          },
          "related_terms": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Related Terms",
            "type": "array"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          },
          "times": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Times",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "vocabulary": {
            "additionalProperties": true,
            "title": "Vocabulary",
            "type": "object"
          }
        },
        "required": [
          "table_id",
          "title",
          "source_url",
          "parse_status",
          "metadata",
          "vocabulary",
          "geographies",
          "times",
          "evidence",
          "related_terms"
        ],
        "title": "TableDetailResponse",
        "type": "object"
      },
      "TermDetailResponse": {
        "description": "Denormalized vocabulary-term detail.",
        "properties": {
          "canonical_term": {
            "title": "Canonical Term",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "cluster": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "occurrence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Occurrence Ids",
            "type": "array"
          },
          "provenance": {
            "additionalProperties": true,
            "title": "Provenance",
            "type": "object"
          },
          "relations": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Relations",
            "type": "object"
          },
          "table_appearances": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Table Appearances",
            "type": "array"
          },
          "term_id": {
            "title": "Term Id",
            "type": "string"
          }
        },
        "required": [
          "term_id",
          "canonical_term"
        ],
        "title": "TermDetailResponse",
        "type": "object"
      },
      "TermListResponse": {
        "description": "Paginated vocabulary-term summaries.",
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Items",
            "type": "array"
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "pagination",
          "items"
        ],
        "title": "TermListResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only APIs over grounded StatVocab artifacts and search indexes.",
    "title": "StatVocab API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/clusters": {
      "get": {
        "description": "Return paginated measure-cluster summaries.",
        "operationId": "clusters_api_clusters_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clusters",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/evaluation": {
      "get": {
        "description": "Return available evaluation summaries.",
        "operationId": "evaluation_api_evaluation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Evaluation",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/health": {
      "get": {
        "description": "Return loaded-run and artifact readiness.",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/relations": {
      "get": {
        "description": "Return paginated measure relations.",
        "operationId": "relations_api_relations_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "term_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Term Id"
            }
          },
          {
            "in": "query",
            "name": "relation_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "broader_than",
                    "narrower_than",
                    "variant_of",
                    "related_to"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Relation Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Relations",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/search": {
      "post": {
        "description": "Search source tables with grounded evidence.",
        "operationId": "search_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search",
        "tags": [
          "search"
        ]
      }
    },
    "/api/tables/{table_id}": {
      "get": {
        "description": "Return one source-table detail payload.",
        "operationId": "table_detail_api_tables__table_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Table Detail",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/terms": {
      "get": {
        "description": "Return paginated vocabulary-term summaries.",
        "operationId": "terms_api_terms_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "measure",
                    "dimension_name",
                    "dimension_value",
                    "unit",
                    "other_ambiguous"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Terms",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/terms/{term_id}": {
      "get": {
        "description": "Return one vocabulary-term detail payload.",
        "operationId": "term_detail_api_terms__term_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Term Detail",
        "tags": [
          "artifacts"
        ]
      }
    }
  }
}